Work around compilation workflow issues
Since we're using the cljs_in_cljs version of the cljsc compiler to
build our code, we have to work around the fact that we don't have
access to proper tooling (like, for instance, lein-cljsbuild) and
fold our dependencies directly into the project. This is a godawful
hack and needs to be addressed more directly in the future.